refactor(pumpkin-core): Cleanup creation of propagator conflict
