+++ pumpkin-crates/core/src/basic_types/mod.rs
@@ -2,7 +2,6 @@ mod constraint_operation_error;
 mod csp_solver_execution_flag;
 mod function;
 mod predicate_id_generators;
-mod propagation_status_cp;
 mod propositional_conjunction;
 mod random;
 mod ref_or_owned;
@@ -18,7 +17,6 @@ pub use function::Function;
 pub(crate) use predicate_id_generators::DeletablePredicateIdGenerator;
 pub use predicate_id_generators::PredicateId;
 pub use predicate_id_generators::PredicateIdGenerator;
-pub use propagation_status_cp::*;
 pub use propositional_conjunction::PropositionalConjunction;
 pub use random::*;
 pub(crate) use ref_or_owned::*;
