+++ pumpkin-crates/core/src/propagators/nogoods/nogood_propagator.rs
@@ -7,13 +7,13 @@ use super::LearningOptions;
 use super::NogoodId;
 use super::NogoodInfo;
 use crate::basic_types::PredicateId;
-use crate::basic_types::PropagationStatusCP;
-use crate::basic_types::PropagatorConflict;
 use crate::basic_types::PropositionalConjunction;
 use crate::containers::KeyedVec;
 use crate::containers::StorageKey;
 use crate::engine::Assignments;
 use crate::engine::Lbd;
+use crate::engine::PropagationStatusCP;
+use crate::engine::PropagatorConflict;
 use crate::engine::notifications::NotificationEngine;
 use crate::engine::predicates::predicate::Predicate;
 use crate::engine::reason::Reason;
