+++ pumpkin-crates/propagators/src/propagators/arithmetic/integer_division.rs
@@ -17,7 +17,7 @@ use pumpkin_core::propagation::Propagator;
 use pumpkin_core::propagation::PropagatorConstructor;
 use pumpkin_core::propagation::PropagatorConstructorContext;
 use pumpkin_core::propagation::ReadDomains;
-use pumpkin_core::results::PropagationStatusCP;
+use pumpkin_core::state::PropagationStatusCP;
 use pumpkin_core::variables::IntegerVariable;
 
 /// The [`PropagatorConstructor`] for the [`DivisionPropagator`].
