+++ pumpkin-crates/propagators/src/propagators/cumulative/time_table/over_interval_incremental_propagator/time_table_over_interval_incremental.rs
@@ -20,9 +20,8 @@ use pumpkin_core::propagation::PropagationContext;
 use pumpkin_core::propagation::Propagator;
 use pumpkin_core::propagation::PropagatorConstructor;
 use pumpkin_core::propagation::PropagatorConstructorContext;
-use pumpkin_core::results::PropagationStatusCP;
-use pumpkin_core::state::Conflict;
-use pumpkin_core::state::PropagatorConflict;
+use pumpkin_core::state::PropagationStatusCP;
+use pumpkin_core::state::propagator_conflict;
 use pumpkin_core::variables::IntegerVariable;
 
 use super::insertion;
@@ -412,10 +411,7 @@ impl<Var: IntegerVariable + 'static, const SYNCHRONISE: bool> Propagator
         );
 
         if self.parameters.is_infeasible {
-            return Err(Conflict::Propagator(PropagatorConflict {
-                conjunction: conjunction!(),
-                inference_code: self.inference_code.clone().unwrap(),
-            }));
+            return propagator_conflict(conjunction!(), self.inference_code.as_ref().unwrap());
         }
 
         self.update_time_table(&mut context)?;
