+++ pumpkin-crates/propagators/src/propagators/cumulative/time_table/per_point_incremental_propagator/synchronisation.rs
@@ -3,8 +3,8 @@ use std::rc::Rc;
 use pumpkin_core::asserts::pumpkin_assert_moderate;
 use pumpkin_core::proof::InferenceCode;
 use pumpkin_core::propagation::Domains;
-use pumpkin_core::results::PropagationStatusCP;
 use pumpkin_core::state::Conflict;
+use pumpkin_core::state::PropagationStatusCP;
 use pumpkin_core::variables::IntegerVariable;
 
 use crate::cumulative::CumulativeParameters;
