+++ pumpkin-crates/propagators/src/propagators/cumulative/time_table/per_point_incremental_propagator/time_table_per_point_incremental.rs
@@ -20,9 +20,8 @@ use pumpkin_core::propagation::PropagationContext;
 use pumpkin_core::propagation::Propagator;
 use pumpkin_core::propagation::PropagatorConstructor;
 use pumpkin_core::propagation::PropagatorConstructorContext;
-use pumpkin_core::results::PropagationStatusCP;
-use pumpkin_core::state::Conflict;
-use pumpkin_core::state::PropagatorConflict;
+use pumpkin_core::state::PropagationStatusCP;
+use pumpkin_core::state::propagator_conflict;
 use pumpkin_core::variables::IntegerVariable;
 
 use crate::cumulative::ArgTask;
@@ -441,10 +440,7 @@ impl<Var: IntegerVariable + 'static + Debug, const SYNCHRONISE: bool> Propagator
         );
 
         if self.parameters.is_infeasible {
-            return Err(Conflict::Propagator(PropagatorConflict {
-                conjunction: conjunction!(),
-                inference_code: self.inference_code.clone().unwrap(),
-            }));
+            return propagator_conflict(conjunction!(), self.inference_code.as_ref().unwrap());
         }
 
         // We update the time-table based on the stored updates
